moving middlware to its own ffile
